@@ -1,7 +1,7 @@
 {
   "name": "bittorrent-tracker",
   "description": "Simple, robust, BitTorrent tracker (client & server) implementation",
-  "version": "9.18.4",
+  "version": "10.0.12",
   "author": {
     "name": "WebTorrent LLC",
     "email": "feross@webtorrent.io",
@@ -25,10 +25,12 @@
   "bugs": {
     "url": "https://github.com/webtorrent/bittorrent-tracker/issues"
   },
+  "type": "module",
   "dependencies": {
-    "bencode": "^2.0.1",
+    "@thaunknown/simple-peer": "^9.12.1",
+    "@thaunknown/simple-websocket": "^9.1.0",
+    "bencode": "^4.0.0",
     "bittorrent-peerid": "^1.3.3",
-    "bn.js": "^5.2.0",
     "chrome-dgram": "^3.0.6",
     "clone": "^2.0.0",
     "compact2string": "^1.4.1",
@@ -39,28 +41,30 @@
     "once": "^1.4.0",
     "queue-microtask": "^1.2.3",
     "random-iterate": "^1.0.1",
-    "randombytes": "^2.1.0",
     "run-parallel": "^1.2.0",
     "run-series": "^1.1.9",
     "simple-get": "^4.0.0",
-    "simple-peer": "^9.11.0",
-    "simple-websocket": "^9.1.0",
     "socks": "^2.0.0",
-    "string2compact": "^1.3.0",
+    "string2compact": "^2.0.0",
+    "uint8-util": "^2.1.9",
     "unordered-array-remove": "^1.0.2",
-    "ws": "^7.4.5"
+    "ws": "^8.0.0"
   },
   "devDependencies": {
-    "@webtorrent/semantic-release-config": "1.0.7",
-    "magnet-uri": "6.2.0",
-    "semantic-release": "18.0.1",
+    "@mapbox/node-pre-gyp": "1.0.11",
+    "@webtorrent/semantic-release-config": "1.0.10",
+    "magnet-uri": "7.0.5",
+    "semantic-release": "21.0.7",
     "standard": "*",
-    "tape": "5.5.2",
-    "webtorrent-fixtures": "1.7.5",
+    "tape": "5.6.6",
+    "webtorrent-fixtures": "2.0.2",
     "wrtc": "0.4.7"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=12.20.0"
+  },
+  "exports": {
+    "import": "./index.js"
   },
   "keywords": [
     "bittorrent",
