@@ -1,14 +1,16 @@
-const arrayRemove = require('unordered-array-remove')
-const bencode = require('bencode')
-const clone = require('clone')
-const compact2string = require('compact2string')
-const debug = require('debug')('bittorrent-tracker:http-tracker')
-const get = require('simple-get')
-const Socks = require('socks')
+import arrayRemove from 'unordered-array-remove'
+import bencode from 'bencode'
+import clone from 'clone'
+import Debug from 'debug'
+import get from 'simple-get'
+import Socks from 'socks'
+import { bin2hex, hex2bin, arr2text, text2arr, arr2hex } from 'uint8-util'
 
-const common = require('../common')
-const Tracker = require('./tracker')
+import common from '../common.js'
+import Tracker from './tracker.js'
+import compact2string from 'compact2string'
 
+const debug = Debug('bittorrent-tracker:http-tracker')
 const HTTP_SCRAPE_SUPPORT = /\/(announce)[^/]*$/
 
 /**
@@ -64,8 +66,8 @@ class HTTPTracker extends Tracker {
     }
 
     const infoHashes = (Array.isArray(opts.infoHash) && opts.infoHash.length > 0)
-      ? opts.infoHash.map(infoHash => infoHash.toString('binary'))
-      : (opts.infoHash && opts.infoHash.toString('binary')) || this.client._infoHashBinary
+      ? opts.infoHash.map(infoHash => hex2bin(infoHash))
+      : (opts.infoHash && hex2bin(opts.infoHash)) || this.client._infoHashBinary
     const params = {
       info_hash: infoHashes
     }
@@ -158,13 +160,13 @@ class HTTPTracker extends Tracker {
       } catch (err) {
         return cb(new Error(`Error decoding tracker response: ${err.message}`))
       }
-      const failure = data['failure reason']
+      const failure = data['failure reason'] && arr2text(data['failure reason'])
       if (failure) {
         debug(`failure from ${requestUrl} (${failure})`)
         return cb(new Error(failure))
       }
 
-      const warning = data['warning message']
+      const warning = data['warning message'] && arr2text(data['warning message'])
       if (warning) {
         debug(`warning from ${requestUrl} (${warning})`)
         self.client.emit('warning', new Error(warning))
@@ -188,15 +190,15 @@ class HTTPTracker extends Tracker {
 
     const response = Object.assign({}, data, {
       announce: this.announceUrl,
-      infoHash: common.binaryToHex(data.info_hash)
+      infoHash: bin2hex(data.info_hash || String(data.info_hash))
     })
     this.client.emit('update', response)
 
     let addrs
-    if (Buffer.isBuffer(data.peers)) {
+    if (ArrayBuffer.isView(data.peers)) {
       // tracker returned compact response
       try {
-        addrs = compact2string.multi(data.peers)
+        addrs = compact2string.multi(Buffer.from(data.peers))
       } catch (err) {
         return this.client.emit('warning', err)
       }
@@ -210,10 +212,10 @@ class HTTPTracker extends Tracker {
       })
     }
 
-    if (Buffer.isBuffer(data.peers6)) {
+    if (ArrayBuffer.isView(data.peers6)) {
       // tracker returned compact response
       try {
-        addrs = compact2string.multi6(data.peers6)
+        addrs = compact2string.multi6(Buffer.from(data.peers6))
       } catch (err) {
         return this.client.emit('warning', err)
       }
@@ -242,12 +244,14 @@ class HTTPTracker extends Tracker {
       return
     }
 
-    keys.forEach(infoHash => {
+    keys.forEach(_infoHash => {
       // TODO: optionally handle data.flags.min_request_interval
       // (separate from announce interval)
-      const response = Object.assign(data[infoHash], {
+      const infoHash = _infoHash.length !== 20 ? arr2hex(text2arr(_infoHash)) : bin2hex(_infoHash)
+
+      const response = Object.assign(data[_infoHash], {
         announce: this.announceUrl,
-        infoHash: common.binaryToHex(infoHash)
+        infoHash
       })
       this.client.emit('scrape', response)
     })
@@ -256,4 +260,4 @@ class HTTPTracker extends Tracker {
 
 HTTPTracker.prototype.DEFAULT_ANNOUNCE_INTERVAL = 30 * 60 * 1000 // 30 minutes
 
-module.exports = HTTPTracker
+export default HTTPTracker
