@@ -1,4 +1,4 @@
-const EventEmitter = require('events')
+import EventEmitter from 'events'
 
 class Tracker extends EventEmitter {
   constructor (client, announceUrl) {
@@ -25,4 +25,4 @@ class Tracker extends EventEmitter {
   }
 }
 
-module.exports = Tracker
+export default Tracker
