@@ -1,14 +1,15 @@
-const arrayRemove = require('unordered-array-remove')
-const BN = require('bn.js')
-const clone = require('clone')
-const compact2string = require('compact2string')
-const debug = require('debug')('bittorrent-tracker:udp-tracker')
-const dgram = require('dgram')
-const randombytes = require('randombytes')
-const Socks = require('socks')
+import arrayRemove from 'unordered-array-remove'
+import clone from 'clone'
+import Debug from 'debug'
+import dgram from 'dgram'
+import Socks from 'socks'
+import { concat, hex2arr, randomBytes } from 'uint8-util'
 
-const common = require('../common')
-const Tracker = require('./tracker')
+import common from '../common.js'
+import Tracker from './tracker.js'
+import compact2string from 'compact2string'
+
+const debug = Debug('bittorrent-tracker:udp-tracker')
 
 /**
  * UDP torrent tracker client (for an individual tracker)
@@ -129,7 +130,7 @@ class UDPTracker extends Tracker {
       }, common.REQUEST_TIMEOUT)
       if (timeout.unref) timeout.unref()
 
-      send(Buffer.concat([
+      send(concat([
         common.CONNECTION_ID,
         common.toUInt32(common.ACTIONS.CONNECT),
         transactionId
@@ -173,7 +174,8 @@ class UDPTracker extends Tracker {
 
     function onSocketMessage (msg) {
       if (proxySocket) msg = msg.slice(10)
-      if (msg.length < 8 || msg.readUInt32BE(4) !== transactionId.readUInt32BE(0)) {
+      const view = new DataView(transactionId.buffer)
+      if (msg.length < 8 || msg.readUInt32BE(4) !== view.getUint32(0)) {
         return onError(new Error('tracker sent invalid transaction id'))
       }
 
@@ -268,14 +270,14 @@ class UDPTracker extends Tracker {
     function announce (connectionId, opts) {
       transactionId = genTransactionId()
 
-      send(Buffer.concat([
+      send(concat([
         connectionId,
         common.toUInt32(common.ACTIONS.ANNOUNCE),
         transactionId,
         self.client._infoHashBuffer,
         self.client._peerIdBuffer,
         toUInt64(opts.downloaded),
-        opts.left != null ? toUInt64(opts.left) : Buffer.from('FFFFFFFFFFFFFFFF', 'hex'),
+        opts.left != null ? toUInt64(opts.left) : hex2arr('ffffffffffffffff'),
         toUInt64(opts.uploaded),
         common.toUInt32(common.EVENTS[opts.event] || 0),
         common.toUInt32(0), // ip address (optional)
@@ -289,10 +291,10 @@ class UDPTracker extends Tracker {
       transactionId = genTransactionId()
 
       const infoHash = (Array.isArray(opts.infoHash) && opts.infoHash.length > 0)
-        ? Buffer.concat(opts.infoHash)
+        ? concat(opts.infoHash)
         : (opts.infoHash || self.client._infoHashBuffer)
 
-      send(Buffer.concat([
+      send(concat([
         connectionId,
         common.toUInt32(common.ACTIONS.SCRAPE),
         transactionId,
@@ -305,12 +307,13 @@ class UDPTracker extends Tracker {
 UDPTracker.prototype.DEFAULT_ANNOUNCE_INTERVAL = 30 * 60 * 1000 // 30 minutes
 
 function genTransactionId () {
-  return randombytes(4)
+  return randomBytes(4)
 }
 
 function toUInt16 (n) {
-  const buf = Buffer.allocUnsafe(2)
-  buf.writeUInt16BE(n, 0)
+  const buf = new Uint8Array(2)
+  const view = new DataView(buf.buffer)
+  view.setUint16(0, n)
   return buf
 }
 
@@ -318,15 +321,14 @@ const MAX_UINT = 4294967295
 
 function toUInt64 (n) {
   if (n > MAX_UINT || typeof n === 'string') {
-    const bytes = new BN(n).toArray()
-    while (bytes.length < 8) {
-      bytes.unshift(0)
-    }
-    return Buffer.from(bytes)
+    const buf = new Uint8Array(8)
+    const view = new DataView(buf.buffer)
+    view.setBigUint64(0, BigInt(n))
+    return buf
   }
-  return Buffer.concat([common.toUInt32(0), common.toUInt32(n)])
+  return concat([new Uint8Array(4), common.toUInt32(n)])
 }
 
 function noop () {}
 
-module.exports = UDPTracker
+export default UDPTracker
