@@ -3,46 +3,49 @@
  * These are separate from common.js so they can be skipped when bundling for the browser.
  */
 
-const querystring = require('querystring')
+import querystring from 'querystring'
+import { concat } from 'uint8-util'
 
-exports.IPV4_RE = /^[\d.]+$/
-exports.IPV6_RE = /^[\da-fA-F:]+$/
-exports.REMOVE_IPV4_MAPPED_IPV6_RE = /^::ffff:/
+export const IPV4_RE = /^[\d.]+$/
+export const IPV6_RE = /^[\da-fA-F:]+$/
+export const REMOVE_IPV4_MAPPED_IPV6_RE = /^::ffff:/
 
-exports.CONNECTION_ID = Buffer.concat([toUInt32(0x417), toUInt32(0x27101980)])
-exports.ACTIONS = { CONNECT: 0, ANNOUNCE: 1, SCRAPE: 2, ERROR: 3 }
-exports.EVENTS = { update: 0, completed: 1, started: 2, stopped: 3 }
-exports.EVENT_IDS = {
+export const CONNECTION_ID = concat([toUInt32(0x417), toUInt32(0x27101980)])
+export const ACTIONS = { CONNECT: 0, ANNOUNCE: 1, SCRAPE: 2, ERROR: 3 }
+export const EVENTS = { update: 0, completed: 1, started: 2, stopped: 3, paused: 4 }
+export const EVENT_IDS = {
   0: 'update',
   1: 'completed',
   2: 'started',
-  3: 'stopped'
+  3: 'stopped',
+  4: 'paused'
 }
-exports.EVENT_NAMES = {
+export const EVENT_NAMES = {
   update: 'update',
   completed: 'complete',
   started: 'start',
-  stopped: 'stop'
+  stopped: 'stop',
+  paused: 'pause'
 }
 
 /**
  * Client request timeout. How long to wait before considering a request to a
  * tracker server to have timed out.
  */
-exports.REQUEST_TIMEOUT = 15000
+export const REQUEST_TIMEOUT = 15000
 
 /**
  * Client destroy timeout. How long to wait before forcibly cleaning up all
  * pending requests, open sockets, etc.
  */
-exports.DESTROY_TIMEOUT = 1000
+export const DESTROY_TIMEOUT = 1000
 
-function toUInt32 (n) {
-  const buf = Buffer.allocUnsafe(4)
-  buf.writeUInt32BE(n, 0)
+export function toUInt32 (n) {
+  const buf = new Uint8Array(4)
+  const view = new DataView(buf.buffer)
+  view.setUint32(0, n)
   return buf
 }
-exports.toUInt32 = toUInt32
 
 /**
  * `querystring.parse` using `unescape` instead of decodeURIComponent, since bittorrent
@@ -50,7 +53,7 @@ exports.toUInt32 = toUInt32
  * @param  {string} q
  * @return {Object}
  */
-exports.querystringParse = q => querystring.parse(q, null, null, { decodeURIComponent: unescape })
+export const querystringParse = q => querystring.parse(q, null, null, { decodeURIComponent: unescape })
 
 /**
  * `querystring.stringify` using `escape` instead of encodeURIComponent, since bittorrent
@@ -58,7 +61,7 @@ exports.querystringParse = q => querystring.parse(q, null, null, { decodeURIComp
  * @param  {Object} obj
  * @return {string}
  */
-exports.querystringStringify = obj => {
+export const querystringStringify = obj => {
   let ret = querystring.stringify(obj, null, null, { encodeURIComponent: escape })
   ret = ret.replace(/[@*/+]/g, char => // `escape` doesn't encode the characters @*/+ so we do it manually
   `%${char.charCodeAt(0).toString(16).toUpperCase()}`)
