@@ -1,8 +1,8 @@
-module.exports = parseHttpRequest
+import { bin2hex } from 'uint8-util'
 
-const common = require('../common')
+import common from '../common.js'
 
-function parseHttpRequest (req, opts) {
+export default function (req, opts) {
   if (!opts) opts = {}
   const s = req.url.split('?')
   const params = common.querystringParse(s[1])
@@ -14,12 +14,12 @@ function parseHttpRequest (req, opts) {
     if (typeof params.info_hash !== 'string' || params.info_hash.length !== 20) {
       throw new Error('invalid info_hash')
     }
-    params.info_hash = common.binaryToHex(params.info_hash)
+    params.info_hash = bin2hex(params.info_hash)
 
     if (typeof params.peer_id !== 'string' || params.peer_id.length !== 20) {
       throw new Error('invalid peer_id')
     }
-    params.peer_id = common.binaryToHex(params.peer_id)
+    params.peer_id = bin2hex(params.peer_id)
 
     params.port = Number(params.port)
     if (!params.port) throw new Error('invalid port')
@@ -33,9 +33,17 @@ function parseHttpRequest (req, opts) {
       common.MAX_ANNOUNCE_PEERS
     )
 
-    params.ip = opts.trustProxy
-      ? req.headers['x-forwarded-for'] || req.connection.remoteAddress
-      : req.connection.remoteAddress.replace(common.REMOVE_IPV4_MAPPED_IPV6_RE, '') // force ipv4
+    if (opts.trustProxy) {
+      if (req.headers['x-forwarded-for']) {
+        const [realIp] = req.headers['x-forwarded-for'].split(',')
+        params.ip = realIp.trim()
+      } else {
+        params.ip = req.connection.remoteAddress
+      }
+    } else {
+      params.ip = req.connection.remoteAddress.replace(common.REMOVE_IPV4_MAPPED_IPV6_RE, '') // force ipv4
+    }
+
     params.addr = `${common.IPV6_RE.test(params.ip) ? `[${params.ip}]` : params.ip}:${params.port}`
 
     params.headers = req.headers
@@ -48,7 +56,7 @@ function parseHttpRequest (req, opts) {
         if (typeof binaryInfoHash !== 'string' || binaryInfoHash.length !== 20) {
           throw new Error('invalid info_hash')
         }
-        return common.binaryToHex(binaryInfoHash)
+        return bin2hex(binaryInfoHash)
       })
     }
   } else {
