@@ -1,8 +1,8 @@
-module.exports = parseWebSocketRequest
+import { bin2hex } from 'uint8-util'
 
-const common = require('../common')
+import common from '../common.js'
 
-function parseWebSocketRequest (socket, opts, params) {
+export default function (socket, opts, params) {
   if (!opts) opts = {}
   params = JSON.parse(params) // may throw
 
@@ -14,18 +14,18 @@ function parseWebSocketRequest (socket, opts, params) {
     if (typeof params.info_hash !== 'string' || params.info_hash.length !== 20) {
       throw new Error('invalid info_hash')
     }
-    params.info_hash = common.binaryToHex(params.info_hash)
+    params.info_hash = bin2hex(params.info_hash)
 
     if (typeof params.peer_id !== 'string' || params.peer_id.length !== 20) {
       throw new Error('invalid peer_id')
     }
-    params.peer_id = common.binaryToHex(params.peer_id)
+    params.peer_id = bin2hex(params.peer_id)
 
     if (params.answer) {
       if (typeof params.to_peer_id !== 'string' || params.to_peer_id.length !== 20) {
         throw new Error('invalid `to_peer_id` (required with `answer`)')
       }
-      params.to_peer_id = common.binaryToHex(params.to_peer_id)
+      params.to_peer_id = bin2hex(params.to_peer_id)
     }
 
     params.left = Number(params.left)
@@ -45,7 +45,7 @@ function parseWebSocketRequest (socket, opts, params) {
         if (typeof binaryInfoHash !== 'string' || binaryInfoHash.length !== 20) {
           throw new Error('invalid info_hash')
         }
-        return common.binaryToHex(binaryInfoHash)
+        return bin2hex(binaryInfoHash)
       })
     }
   } else {
@@ -55,9 +55,17 @@ function parseWebSocketRequest (socket, opts, params) {
   // On first parse, save important data from `socket.upgradeReq` and delete it
   // to reduce memory usage.
   if (socket.upgradeReq) {
-    socket.ip = opts.trustProxy
-      ? socket.upgradeReq.headers['x-forwarded-for'] || socket.upgradeReq.connection.remoteAddress
-      : socket.upgradeReq.connection.remoteAddress.replace(common.REMOVE_IPV4_MAPPED_IPV6_RE, '') // force ipv4
+    if (opts.trustProxy) {
+      if (socket.upgradeReq.headers['x-forwarded-for']) {
+        const [realIp] = socket.upgradeReq.headers['x-forwarded-for'].split(',')
+        socket.ip = realIp.trim()
+      } else {
+        socket.ip = socket.upgradeReq.connection.remoteAddress
+      }
+    } else {
+      socket.ip = socket.upgradeReq.connection.remoteAddress.replace(common.REMOVE_IPV4_MAPPED_IPV6_RE, '') // force ipv4
+    }
+
     socket.port = socket.upgradeReq.connection.remotePort
     if (socket.port) {
       socket.addr = `${common.IPV6_RE.test(socket.ip) ? `[${socket.ip}]` : socket.ip}:${socket.port}`
