@@ -1,7 +1,9 @@
-const arrayRemove = require('unordered-array-remove')
-const debug = require('debug')('bittorrent-tracker:swarm')
-const LRU = require('lru')
-const randomIterate = require('random-iterate')
+import arrayRemove from 'unordered-array-remove'
+import Debug from 'debug'
+import LRU from 'lru'
+import randomIterate from 'random-iterate'
+
+const debug = Debug('bittorrent-tracker:swarm')
 
 // Regard this as the default implementation of an interface that you
 // need to support when overriding Server.createSwarm() and Server.getSwarm()
@@ -47,6 +49,8 @@ class Swarm {
       self._onAnnounceCompleted(params, peer, id)
     } else if (params.event === 'update') {
       self._onAnnounceUpdate(params, peer, id)
+    } else if (params.event === 'paused') {
+      self._onAnnouncePaused(params, peer, id)
     } else {
       cb(new Error('invalid event'))
       return
@@ -132,6 +136,15 @@ class Swarm {
     this.peers.set(id, peer)
   }
 
+  _onAnnouncePaused (params, peer, id) {
+    if (!peer) {
+      debug('unexpected `paused` event from peer that is not in swarm')
+      return this._onAnnounceStarted(params, peer, id) // treat as a start
+    }
+
+    this._onAnnounceUpdate(params, peer, id)
+  }
+
   _getPeers (numwant, ownPeerId, isWebRTC) {
     const peers = []
     const ite = randomIterate(this.peers.keys)
@@ -148,4 +161,4 @@ class Swarm {
   }
 }
 
-module.exports = Swarm
+export default Swarm
