@@ -1,7 +1,7 @@
-const Client = require('../')
-const common = require('./common')
-const fixtures = require('webtorrent-fixtures')
-const test = require('tape')
+import Client from '../index.js'
+import common from './common.js'
+import fixtures from 'webtorrent-fixtures'
+import test from 'tape'
 
 const peerId = Buffer.from('01234567890123456789')
 const port = 6881
