@@ -1,23 +1,11 @@
 /**
  * Functions/constants needed by both the client and server.
  */
+import * as common from './common-node.js'
+export * from './common-node.js'
 
-exports.DEFAULT_ANNOUNCE_PEERS = 50
-exports.MAX_ANNOUNCE_PEERS = 82
-
-exports.binaryToHex = str => {
-  if (typeof str !== 'string') {
-    str = String(str)
-  }
-  return Buffer.from(str, 'binary').toString('hex')
-}
-
-exports.hexToBinary = str => {
-  if (typeof str !== 'string') {
-    str = String(str)
-  }
-  return Buffer.from(str, 'hex').toString('binary')
-}
+export const DEFAULT_ANNOUNCE_PEERS = 50
+export const MAX_ANNOUNCE_PEERS = 82
 
 // HACK: Fix for WHATWG URL object not parsing non-standard URL schemes like
 // 'udp:'. Just replace it with 'http:' since we only need a few properties.
@@ -31,7 +19,7 @@ exports.hexToBinary = str => {
 // Bug reports:
 // - Chrome: https://bugs.chromium.org/p/chromium/issues/detail?id=734880
 // - Firefox: https://bugzilla.mozilla.org/show_bug.cgi?id=1374505
-exports.parseUrl = str => {
+export const parseUrl = str => {
   const url = new URL(str.replace(/^udp:/, 'http:'))
 
   if (str.match(/^udp:/)) {
@@ -45,5 +33,9 @@ exports.parseUrl = str => {
   return url
 }
 
-const config = require('./common-node')
-Object.assign(exports, config)
+export default {
+  DEFAULT_ANNOUNCE_PEERS,
+  MAX_ANNOUNCE_PEERS,
+  parseUrl,
+  ...common
+}
