@@ -1,9 +1,8 @@
-module.exports = parseUdpRequest
+import ipLib from 'ip'
+import common from '../common.js'
+import { equal } from 'uint8-util'
 
-const ipLib = require('ip')
-const common = require('../common')
-
-function parseUdpRequest (msg, rinfo) {
+export default function (msg, rinfo) {
   if (msg.length < 16) throw new Error('received packet is too short')
 
   const params = {
@@ -13,7 +12,7 @@ function parseUdpRequest (msg, rinfo) {
     type: 'udp'
   }
 
-  if (!common.CONNECTION_ID.equals(params.connectionId)) {
+  if (!equal(common.CONNECTION_ID, params.connectionId)) {
     throw new Error('received packet with invalid connection id')
   }
 
